requirements.txt
Flask==2.3.2
requests==2.26.0
deepgram-sdk==1.1.0
google-ai-generativelanguage==0.2.0
python-dotenv==1.0.1
httpx==0.23.3
